uvicornclient/requirements.txt
grpcio
fastapi
protobuf
uvicorn
